Edited class and testing files
Edited class sent by Hadeel in assignment zip.
However there are problems with reading the file (i.e. it just doesn't
read it), didn't implement Dijkstra's algorithm yet so one of you guys
do it (it's in the slides), and I didn't implement an algorithm for part
two yet (it's either really easy or it's going to be really annoying)